+++ pyswtools/stat/main.py
@@ -3,27 +3,19 @@ Give stat from an assembly
 """
 
 import os
-from enum import Enum
 import click
 
 # pylint: disable=relative-beyond-top-level
 from ..utils import check_system, check_system_verbose
-from ..helper_sw import open_app_and_file, is_temp, is_assembly
+from ..helper_sw import open_app_and_file, is_temp, is_assembly, open_drawing
 
-
-class TypeOutput(str, Enum):
-    """Class represeting an output type"""
-
-    TREE = "tree"
-    LIST = "list"
-
-
-class TypeSort(str, Enum):
-    """Class represeting an output type"""
-
-    MASS = "mass"
-    MASS_PART = "mass-part"
-    NAME = "name"
+from .definitions import (
+    TypeComponent,
+    TypeOutput,
+    TypeSort,
+    StatComponent,
+    StatComponentTree,
+)
 
 
 if check_system():
@@ -40,7 +32,7 @@ def filter_density_list(struct: dict) -> None:
     """
     Filter element from a list struct to get only the ones with a density differnt from 1000
     """
-    return {k: v for k, v in struct.items() if abs(v["density"] - 1000) < 1e-3}
+    return {k: v for k, v in struct.items() if abs(v.density - 1000) < 1e-3}
 
 
 def filter_density_tree(struct: dict) -> None:
@@ -49,10 +41,50 @@ def filter_density_tree(struct: dict) -> None:
     """
     return_struct = {}
     for k, v in struct.items():
-        if (v["density"] - 1000) < 1e-4:
+        if (v.density - 1000) < 1e-4:
             return_struct[k] = v
-            filter_density_tree(return_struct[k]["children"])
+            filter_density_tree(return_struct[k].children)
     struct = return_struct
+
+
+def filter_component_type(
+    tree_struct: dict, list_struct: dict, type_component=TypeComponent.ALL
+) -> None:
+    """
+    Filter element from a list and tree struct to get one the ones with the matching TypeComponent
+    """
+    if type_component.value == TypeComponent.ALL.value:
+        return
+
+    if type_component.value == TypeComponent.ASSEMBLY.value:
+        filter_component_type_tree(tree_struct, list_struct)
+
+    for k, v in list(list_struct.items()):
+        if v.typeComponent.value != type_component.value:
+            list_struct.pop(k)
+
+
+def filter_component_type_tree(tree_struct: dict, list_struct: dict) -> None:
+    """
+    Filter elements from a list
+    """
+    for k, v in list(tree_struct.items()):
+        if list_struct[k].typeComponent.value == TypeComponent.PART.value:
+            tree_struct.pop(k)
+        else:
+            filter_component_type_tree(v.children, list_struct)
+
+
+def clean_confs(tree_struct: dict, dict_struct: dict) -> None:
+    """
+    Clean the name of the tree and list struct of the conf and remove duplicate.
+    """
+
+    remove_duplicate_conf_list(dict_struct, dict_struct)
+    remove_duplicate_conf_tree(tree_struct, dict_struct)
+
+    remove_conf_list(dict_struct)
+    remove_conf_tree(tree_struct, dict_struct)
 
 
 def remove_conf_tree(tree_struct: dict, dict_of_comp: dict) -> None:
@@ -63,7 +95,7 @@ def remove_conf_tree(tree_struct: dict, dict_of_comp: dict) -> None:
     """
     remove_conf_with_list(tree_struct, dict_of_comp)
     for elem in tree_struct:
-        remove_conf_tree(tree_struct[elem]["children"], dict_of_comp)
+        remove_conf_tree(tree_struct[elem].children, dict_of_comp)
 
 
 def remove_conf_with_list(list_struct: dict, dict_of_comp: dict) -> None:
@@ -77,12 +109,16 @@ def remove_conf_with_list(list_struct: dict, dict_of_comp: dict) -> None:
     list_of_name = list(list_struct.keys())
 
     for name in list_of_name:
-        if strip_conf(name) in dict_of_comp:
-            list_struct[strip_conf(name)] = list_struct[name]
+        stripped_name = strip_conf(name)
+        if stripped_name in dict_of_comp:
+            if stripped_name not in list_struct:
+                list_struct[stripped_name] = list_struct[name]
+            else:
+                list_struct[stripped_name].number += list_struct[name].number
             del list_struct[name]
 
 
-def remove_conf(list_struct: dict) -> None:
+def remove_conf_list(list_struct: dict) -> None:
     """
     Clean the name of a list struct.
     If there is only one configuration for a component, remove the configuration
@@ -122,7 +158,17 @@ def remove_conf(list_struct: dict) -> None:
             del list_struct[name]
 
 
-def remove_duplicate_conf(struct: dict) -> None:
+def remove_duplicate_conf_tree(tree_struct: dict, dict_struct: dict) -> None:
+    """
+    Filter element from struct (tree or list) which have only different configuration names but same Mass
+    """
+
+    remove_conf_with_list(tree_struct, dict_struct)
+    for name in list(tree_struct.keys()):
+        remove_duplicate_conf_tree(tree_struct[name].children, dict_struct)
+
+
+def remove_duplicate_conf_list(struct: dict, mass_struct: dict) -> None:
     """
     Filter element from struct (tree or list) which have only different configuration names but same Mass
     """
@@ -137,15 +183,13 @@ def remove_duplicate_conf(struct: dict) -> None:
         # We test the following elements
         for name_test in sorted_list[idx + 1 :]:
             # if we do not have the same mass, we end
-            if abs(struct[name]["mass"] - struct[name_test]["mass"]) > 1e-5:
+            if abs(mass_struct[name].mass - mass_struct[name_test].mass) > 1e-5:
                 break
             # We compare the names
             if strip_conf(name_test) == strip_conf(name):
                 # We fused data
-                struct[name]["number"] += struct[name_test]["number"]
+                struct[name].number += struct[name_test].number
                 del struct[name_test]
-        if "children" in struct[name]:
-            remove_duplicate_conf(struct[name]["children"])
 
 
 def sort_key_struct(struct: dict, type_sort: TypeSort = TypeSort.MASS) -> list:
@@ -157,35 +201,62 @@ def sort_key_struct(struct: dict, type_sort: TypeSort = TypeSort.MASS) -> list:
     if type_sort is TypeSort.MASS:
         return sorted(
             struct.keys(),
-            key=lambda i: struct[i]["mass"] * struct[i]["number"],
+            key=lambda i: struct[i].mass * struct[i].number,
             reverse=True,
         )
     if type_sort is TypeSort.MASS_PART:
         return sorted(
             struct.keys(),
-            key=lambda i: struct[i]["mass"],
+            key=lambda i: struct[i].mass,
             reverse=True,
         )
 
     return struct.keys()
 
 
+def sort_key_struct_tree(
+    tree_struct: dict, dict_struct: dict, type_sort: TypeSort = TypeSort.MASS
+) -> list:
+    """
+    Sort a dict struct following a given sort
+    """
+    if type_sort is TypeSort.NAME:
+        return sorted(tree_struct.keys(), key=str.lower)
+    if type_sort is TypeSort.MASS:
+        return sorted(
+            tree_struct.keys(),
+            key=lambda i: dict_struct[i].mass * tree_struct[i].number,
+            reverse=True,
+        )
+    if type_sort is TypeSort.MASS_PART:
+        return sorted(
+            tree_struct.keys(),
+            key=lambda i: dict_struct[i].mass,
+            reverse=True,
+        )
+
+    return tree_struct.keys()
+
+
 def print_header():
     """
     Display the header of the output
     """
-    header = f"{'Name':<50} | {'Mtot':<5} | {'n':<3} | {'Mpart':<6} | {'Density':<7}  "
+    header = f"{'Name':<50} | {'Mtot':<5} | {'n':<3} | {'Mpart':<6} | {'Density':<7} | {'Comp':<4} | {'Drw':<3}"
     click.echo(header)
     click.echo("-" * len(header))
 
 
 def display_tree(
-    tree_struct_mass: dict, type_sort: TypeSort = TypeSort.NAME, indent: str = ""
+    tree_struct_mass: dict,
+    dict_struct_mass: dict,
+    type_sort: TypeSort = TypeSort.NAME,
+    indent: str = "",
 ) -> None:
     """
     Display a tree struct recursively
     """
-    sorted_list = sort_key_struct(tree_struct_mass, type_sort)
+    sorted_list = sort_key_struct_tree(tree_struct_mass, dict_struct_mass, type_sort)
     n = len(sorted_list)
     if indent == "":
         print_header()
@@ -194,13 +265,18 @@ def display_tree(
         if idx == n - 1:
             char = "┕━"
         click.echo(
-            f"{indent + char + ' ' + elem :<50} | {tree_struct_mass[elem]['mass'] * tree_struct_mass[elem]['number']:5.3f} | {tree_struct_mass[elem]['number']:3d} | {tree_struct_mass[elem]['mass']:6.4f} | {tree_struct_mass[elem]['density']:6.4f}"
+            f"{indent + char + ' ' + elem :<50} | {dict_struct_mass[elem].mass * tree_struct_mass[elem].number:5.3f} | {tree_struct_mass[elem].number:3d} | {dict_struct_mass[elem].mass:6.4f} | {dict_struct_mass[elem].density:6.2f} | {'Part' if dict_struct_mass[elem].typeComponent == TypeComponent.PART else 'Ass.'} | {dict_struct_mass[elem].numberDrawing:3d}"
         )
-        if len(tree_struct_mass[elem]["children"]) > 0:
+        if len(tree_struct_mass[elem].children) > 0:
             char = "│ "
             if idx == n - 1:
                 char = "  "
-            display_tree(tree_struct_mass[elem]["children"], type_sort, indent + char)
+            display_tree(
+                tree_struct_mass[elem].children,
+                dict_struct_mass,
+                type_sort,
+                indent + char,
+            )
 
 
 def display_list(list_struct: dict, type_sort: TypeSort = TypeSort.NAME) -> None:
@@ -210,7 +286,7 @@ def display_list(list_struct: dict, type_sort: TypeSort = TypeSort.NAME) -> None
     print_header()
     for elem in sort_key_struct(list_struct, type_sort):
         click.echo(
-            f"{'- ' + elem :<50} | {list_struct[elem]['mass'] * list_struct[elem]['number']:5.3f} | {list_struct[elem]['number']:3d} | {list_struct[elem]['mass']:6.4f} | {list_struct[elem]['density']:6.4f}"
+            f"{'- ' + elem :<50} | {list_struct[elem].mass * list_struct[elem].number:5.3f} | {list_struct[elem].number:3d} | {list_struct[elem].mass:6.4f} | {list_struct[elem].density:6.2f} | {'Part' if list_struct[elem].typeComponent == TypeComponent.PART else 'Ass.'} | {list_struct[elem].numberDrawing:3d}"
         )
 
 
@@ -234,6 +310,34 @@ def strip_conf(name: str) -> str:
     return name.split("@")[0]
 
 
+def fill_drawing_dependencies(sw_app, input_path: str, struct: dict):
+    """
+    Get the number of drawings related to each parts
+    """
+    list_of_drawings = []
+    for r, _, fs in os.walk(os.path.dirname(input_path)):
+        for f in fs:
+            if "SLDDRW" in f and "~$" not in f:
+                list_of_drawings.append(os.path.join(r, f))
+
+    parts_in_plan_dict = {}
+    for drawing_path in list_of_drawings:
+        parts_in_plan = sw_app.GetDocumentDependencies2(
+            os.path.abspath(drawing_path), False, True, False
+        )
+        for part_in_plan in parts_in_plan[::2]:
+            if part_in_plan not in parts_in_plan_dict:
+                parts_in_plan_dict[part_in_plan] = 1
+            else:
+                parts_in_plan_dict[part_in_plan] += 1
+
+    for name in struct.keys():
+        if strip_conf(name) in parts_in_plan_dict:
+            struct[name].numberDrawing = parts_in_plan_dict[strip_conf(name)]
+    # TODO: handle configurations
+    return parts_in_plan_dict
+
+
 def complete_info_on_list(sw_comp_children, dict_of_comp: dict) -> dict:
     """
     Get all the info from a list of component
@@ -254,7 +358,7 @@ def complete_info_on_list(sw_comp_children, dict_of_comp: dict) -> dict:
         if sw_child_name not in children:
             children[sw_child_name] = child
         else:
-            children[sw_child_name]["number"] += 1
+            children[sw_child_name].number += 1
 
     return children
 
@@ -269,7 +373,7 @@ def complete_info_assembly(sw_comp, dict_of_comp: dict) -> dict:
 
     # If already there only increase the number
     if sw_comp_name in dict_of_comp:
-        dict_of_comp[sw_comp_name]["number"] += 1
+        dict_of_comp[sw_comp_name].number += 1
     else:
         sw_mass = 0
         sw_density = 0
@@ -288,23 +392,27 @@ def complete_info_assembly(sw_comp, dict_of_comp: dict) -> dict:
             click.echo(f"Could not evaluate {sw_comp_name}")
 
         # Create an new entity in the general dict
-        dict_of_comp[sw_comp_name] = {
-            "number": 1,
-            "mass": sw_mass,
-            "density": sw_density,
-        }
+        dict_of_comp[sw_comp_name] = StatComponent(
+            mass=sw_mass,
+            density=sw_density,
+            number=1,
+            typeComponent=(
+                TypeComponent.ASSEMBLY
+                if len(sw_comp.GetChildren) > 0
+                else TypeComponent.PART
+            ),
+            numberDrawing=0,
+        )
 
     # Get info about children
     sw_comp_children = sw_comp.GetChildren
     children = complete_info_on_list(sw_comp_children, dict_of_comp)
 
     # Create the entry in the tree dict
-    return {
-        "mass": dict_of_comp[sw_comp_name]["mass"],
-        "density": dict_of_comp[sw_comp_name]["density"],
-        "number": 1,
-        "children": children,
-    }
+    return StatComponentTree(
+        number=1,
+        children=children,
+    )
 
 
 @click.command()
@@ -325,11 +433,18 @@ def complete_info_assembly(sw_comp, dict_of_comp: dict) -> dict:
 @click.option(
     "--type-sort", "type_sort", type=click.Choice(TypeSort), default=TypeSort.MASS
 )
+@click.option(
+    "--filter",
+    "type_component",
+    type=click.Choice(TypeComponent),
+    default=TypeComponent.ALL,
+)
 @click.option("--only-default-density", "only_default_density", is_flag=True)
 def stat(
     input_path: str,
     type_output: TypeOutput,
     type_sort: TypeSort,
+    type_component: TypeComponent,
     only_default_density: bool,
 ) -> None:
     """
@@ -343,7 +458,7 @@ def stat(
         click.echo(f"{input_path} is not an assembly file")
         return
 
-    _, sw_doc, filename = open_app_and_file(input_path)
+    sw_app, sw_doc, filename = open_app_and_file(input_path)
 
     # Get list of components
     sw_comps = sw_doc.GetComponents(True)
@@ -351,29 +466,34 @@ def stat(
     # Empty struct to fill
     assembly_name = filename.rpartition(".")[-0] + "@" + sw_doc.GetConfigurationNames[0]
     dict_of_comp = {
-        f"{assembly_name}": {
-            "mass": sw_doc.Extension.CreateMassProperty2.Mass,
-            "density": sw_doc.Extension.CreateMassProperty2.Density,
-            "number": 1,
-        }
+        assembly_name: StatComponent(
+            mass=sw_doc.Extension.CreateMassProperty2.Mass,
+            density=sw_doc.Extension.CreateMassProperty2.Density,
+            number=1,
+            typeComponent=TypeComponent.ASSEMBLY,
+            numberDrawing=0,
+        )
     }
 
     tree_of_comp = {
-        f"{assembly_name}": {
-            "mass": dict_of_comp[assembly_name]["mass"],
-            "density": dict_of_comp[assembly_name]["density"],
-            "number": 1,
-            "children": complete_info_on_list(sw_comps, dict_of_comp),
-        }
+        assembly_name: StatComponentTree(
+            number=1,
+            children=complete_info_on_list(sw_comps, dict_of_comp),
+        )
     }
 
-    remove_duplicate_conf(dict_of_comp)
-    remove_conf(dict_of_comp)
+    fill_drawing_dependencies(sw_app, input_path, dict_of_comp)
+
+    clean_confs(tree_of_comp, dict_of_comp)
+
+    print(tree_of_comp)
+
+    filter_component_type(tree_of_comp, dict_of_comp, type_component)
+
     if type_output is TypeOutput.TREE:
-        remove_duplicate_conf(tree_of_comp)
-        remove_conf_tree(tree_of_comp, dict_of_comp)
-        display_tree(tree_of_comp, type_sort)
+        display_tree(tree_of_comp, dict_of_comp, type_sort)
     elif type_output is TypeOutput.LIST:
+
         if only_default_density:
             dict_of_comp = filter_density_list(dict_of_comp)
         display_list(dict_of_comp, type_sort)